--- main.js
@@ -1,2 +1,0 @@
-module.exports=require('./coerce.js');
-require('./coerce-standard.js');
