+++ test/01 - The basics/01 - types.js
@@ -1,6 +1,7 @@
 var assert = require('chai').assert;
 
-var api = require('../../main.js');
+var api = require('../../coerce.js');
+require('../../coerce-datatypes.js');
 
 describe('Type coercion', function () {
 	it('string -> number', function () {
